pyproject.toml
[project]
name = "luxonis-task"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Martin Zaťko", email = "zatko.m@pm.me"},
]
dependencies = [
    "dependency-injector-fork>=4.42.1",
    "psycopg[binary]>=3.1.18",
    "scrapy>=2.11.1",
    "sqlalchemy>=2.0.27",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
